Split ToolController into model and gui.
